Updated flow
This is the current flow I have been working on for a Node-red GUI that works with Raspberry Pi compute modules in Temco products.  It is currently a work in progress, many features like multiple rooms/zones are currently disabled - I will be finishing up a major overhaul to it within a few days and then will post a more updated working version.